Chromatographic Data Backend for Spectra Objects

The ChromBackendSpectra class extends ChromBackendMemory, inheriting all its slots and methods while providing additional functionality for summarizing chromatographic data from Spectra::Spectra() objects.

It can be initialized with a Spectra object, which is stored in the spectra slot of the backend. Users can also provide a data.frame containing chromatographic metadata, stored in chromData. This metadata filters the Spectra object and generates peaksData. If chromData is not provided, a default data.frame is created from the Spectra data. An "rtmin", "rtmax", "mzmin", and "mzmax" column will be created by condensing the Spectra data corresponding to each unique combination of the factorize.by variables.

The dataOrigin core chromatogram variable should reflect the dataOrigin of the Spectra object. The factorize.by parameter defines the variables for grouping Spectra data into chromatographic data. The default is c("msLevel", "dataOrigin"), which will define separate chromatograms for each combination of msLevel and dataOrigin. These variables must be in both Spectra and chromData (if provided).

The summarize.method parameter defines how spectral data intensity is summarized:

"sum": Sums intensity to create a Total Ion Chromatogram (TIC).

"max": Takes max intensity for a Base Peak Chromatogram (BPC).

If chromData or its factorization columns are modified, the factorize() method must be called to update chromSpectraIndex.

ChromBackendSpectra()

# S4 method for class 'ChromBackendSpectra'
backendInitialize(
  object,
  spectra = Spectra::Spectra(),
  factorize.by = c("msLevel", "dataOrigin"),
  summarize.method = c("sum", "max"),
  chromData = fillCoreChromVariables(),
  ...
)

chromSpectraIndex(object)

Arguments

object

A ChromBackendSpectra object.

spectra

A Spectra object.

factorize.by

A character vector of variables for grouping Spectra data into chromatographic data. Default: c("msLevel", "dataOrigin"). If chromData is provided, it must contain these columns.

summarize.method

A character string specifying intensity summary: "sum" (default) or "max".

chromData

A data.frame with chromatographic data for use in backendInitialize(). If missing, a default is generated. Columns like rtmin, rtmax, mzmin, and mzmax must be provided and not contain NA values. Use -Inf/Inf for unspecified values. The "dataOrigin" column must match the Spectra object's "dataOrigin".

...

Additional parameters.

Value

Refer to the individual function description for information on the return value.

Details

No peaksData is stored until the user calls a function that generates it (e.g., rtime(), peaksData(), intensity()). The peaksData slot replacement is unsupported — modifications are temporary to optimize memory. The inMemory slot indicates this with TRUE.

ChromBackendSpectra should reuse ChromBackendMemory methods whenever possible to keep implementations simple.

Note

ensure that it returns a factor

Author

Philippine Louail, Johannes Rainer.

Examples

library(Spectra)
#> Loading required package: S4Vectors
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> Loading required package: generics
#> 
#> Attaching package: ‘generics’
#> The following objects are masked from ‘package:base’:
#> 
#>     as.difftime, as.factor, as.ordered, intersect, is.element, setdiff,
#>     setequal, union
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, is.unsorted, lapply,
#>     mapply, match, mget, order, paste, pmax, pmax.int, pmin, pmin.int,
#>     rank, rbind, rownames, sapply, saveRDS, table, tapply, unique,
#>     unsplit, which.max, which.min
#> 
#> Attaching package: ‘S4Vectors’
#> The following object is masked from ‘package:utils’:
#> 
#>     findMatches
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> 
#> Attaching package: ‘Spectra’
#> The following object is masked from ‘package:Chromatograms’:
#> 
#>     reset
library(MsBackendMetaboLights)

## Get Spectra data from MetaboLights
be <- backendInitialize(MsBackendMetaboLights(),
    mtblsId = "MTBLS39",
    filePattern = c("63B.cdf")
)
#> Used data files from the assay's column "Raw Spectral Data File" since none were available in column "Derived Spectral Data File".
s <- Spectra(be)

## Initialize ChromBackendSpectra
be_empty <- new("ChromBackendSpectra")
be <- backendInitialize(be_empty, s)

## replace the msLevel data
msLevel(be) <- c(1L, 2L, 3L)
#> Please keep in mind the 'ChromBackendSpectra' backend is read-only. The chromData slot will be modified but the changes will not affect the Spectra object. You will need to run `factorize()` to update the 'chromSpectraIndex' column.

## re-factorize the data
be <- factorize(be)

## Create BPC : we summarize the intensity present in the Spectra object
## by the maximum value, thus creating a Base Peak Chromatogram.
be <- backendInitialize(be_empty, s, summarize.method = "max")

## Can now see the details of this bpc by looking at the chromData of our
## object
chromData(be)
#>   msLevel    rtmin    rtmax mzMin mzMax  mz polarity
#> 1       1 0.296192 2695.833  -Inf   Inf Inf       -1
#> 2       1 0.295936 2697.736  -Inf   Inf Inf       -1
#> 3       1 3.660864 2698.993  -Inf   Inf Inf       -1
#>                                       dataOrigin
#> 1 /github/home/.cache/R/BiocFileCache/AM063B.cdf
#> 2 /github/home/.cache/R/BiocFileCache/CS063B.cdf
#> 3 /github/home/.cache/R/BiocFileCache/MN063B.cdf
#>                                  chromSpectraIndex chromIndex collisionEnergy
#> 1 1_/github/home/.cache/R/BiocFileCache/AM063B.cdf          1              NA
#> 2 1_/github/home/.cache/R/BiocFileCache/CS063B.cdf          2              NA
#> 3 1_/github/home/.cache/R/BiocFileCache/MN063B.cdf          3              NA
#>   precursorMz precursorMzMin precursorMzMax productMz productMzMin productMzMax
#> 1          NA             NA             NA        NA           NA           NA
#> 2          NA             NA             NA        NA           NA           NA
#> 3          NA             NA             NA        NA           NA           NA

## Another possibilities is to create eics from the Spectra object.
## Here we create an EIC with a specific m/z and retention time window.
df <- data.frame(mzmin = 100.01, mzmax = 100.02 , rtmin = 50, rtmax = 100)
be <- backendInitialize(be_empty, s, summarize.method = "sum")
chromData(be) <- cbind(chromData(be), df)
#> Please keep in mind the 'ChromBackendSpectra' backend is read-only. The chromData slot will be modified but the changes will not affect the Spectra object. You will need to run `factorize()` to update the 'chromSpectraIndex' column.

## now when we call the peaksData function, we will get the intensity
## of the spectra object that are in the m/z and retention time window
## defined in the chromData.
peaksData(be)
#> [[1]]
#>           rtime  intensity
#> 1      0.296192   30360.34
#> 2      6.195200   31987.51
#> 3     12.064960   29799.62
#> 4     17.966720   30340.36
#> 5     23.828864   29311.18
#> 6     29.704960   29652.89
#> 7     35.582848   29766.87
#> 8     41.444736   30578.23
#> 9     47.290688   29960.14
#> 10    53.146816   28774.21
#> 11    58.994752   29589.01
#> 12    64.882560   30948.68
#> 13    70.766336   30950.94
#> 14    76.644160   30605.69
#> 15    82.478080   30008.49
#> 16    88.318272   30009.77
#> 17    94.202560   30647.52
#> 18   100.074368   30129.05
#> 19   105.960320   30153.38
#> 20   111.822528   31059.49
#> 21   117.682176   30620.76
#> 22   123.480128   36482.01
#> 23   128.802624  222100.37
#> 24   133.473856  688834.19
#> 25   137.584448  971540.01
#> 26   141.710976  956108.32
#> 27   145.878336  839189.47
#> 28   150.031616  759789.46
#> 29   154.198016  876450.28
#> 30   158.358720 1011332.66
#> 31   162.483072 1044635.20
#> 32   166.643648 1039582.92
#> 33   170.806080  979659.65
#> 34   174.948096  904107.69
#> 35   179.124224  833332.26
#> 36   183.335168  668250.96
#> 37   187.540096  551291.45
#> 38   191.809216  435746.81
#> 39   196.109696  344413.73
#> 40   200.418688  283672.37
#> 41   204.763520  240485.01
#> 42   209.126912  220168.52
#> 43   213.506560  206024.20
#> 44   217.893504  199993.81
#> 45   222.266560  195714.69
#> 46   226.668224  196889.42
#> 47   231.028352  201073.82
#> 48   235.408640  213153.71
#> 49   239.775808  221417.68
#> 50   244.097920  220917.06
#> 51   248.457280  224704.87
#> 52   252.832192  242836.63
#> 53   257.183872  252394.28
#> 54   261.556928  261487.87
#> 55   265.891328  256380.39
#> 56   270.253248  266472.74
#> 57   274.645696  258453.41
#> 58   279.042240  245335.90
#> 59   283.490560  227658.31
#> 60   287.890112  239763.07
#> 61   292.230336  332453.76
#> 62   296.513664  422230.96
#> 63   300.733888  492924.57
#> 64   304.979712  505369.47
#> 65   309.227520  510875.81
#> 66   313.462272  511342.45
#> 67   317.706688  520132.84
#> 68   321.966784  552673.46
#> 69   326.214080  570973.27
#> 70   330.460736  548095.26
#> 71   334.686976  501242.60
#> 72   338.958464  436004.47
#> 73   343.245248  388533.44
#> 74   347.551296  333392.57
#> 75   351.931776  281338.94
#> 76   356.356928  236364.04
#> 77   360.828096  191468.96
#> 78   365.442496  152737.84
#> 79   370.056256  131332.69
#> 80   374.767168  113620.81
#> 81   379.500736  104863.34
#> 82   384.296576   95631.89
#> 83   389.071680   91042.62
#> 84   393.910656   91409.82
#> 85   398.654720   97696.88
#> 86   403.379328   98815.16
#> 87   408.100096   97598.44
#> 88   412.849664   96996.69
#> 89   417.619008   93436.00
#> 90   422.440064   92314.49
#> 91   427.294016   89806.90
#> 92   432.137792   88945.44
#> 93   436.988544   86354.81
#> 94   441.809536   94686.14
#> 95   446.624896   96717.85
#> 96   451.438336   94759.20
#> 97   456.292032   94509.35
#> 98   461.093888   96046.39
#> 99   465.911232  100077.85
#> 100  470.644864  107069.35
#> 101  475.378368  110390.71
#> 102  480.063872  118316.82
#> 103  484.777792  121043.96
#> 104  489.459904  123826.29
#> 105  494.173696  118651.90
#> 106  498.861760  118376.09
#> 107  503.577152  119233.56
#> 108  508.269440  120349.11
#> 109  512.995008  115246.83
#> 110  517.736448  106353.55
#> 111  522.567360   96481.13
#> 112  527.413760   90212.10
#> 113  532.333824   83610.88
#> 114  537.291776   78815.33
#> 115  542.289856   71762.48
#> 116  547.336960   70598.69
#> 117  552.320768   69766.48
#> 118  557.345344   67403.57
#> 119  562.348480   65967.64
#> 120  567.369984   62600.18
#> 121  572.457088   61745.27
#> 122  577.527360   59603.50
#> 123  582.657408   56082.44
#> 124  587.847168   53610.79
#> 125  593.019456   53760.69
#> 126  598.183296   52473.53
#> 127  603.329408   54639.22
#> 128  608.443072   55891.05
#> 129  613.555072   55062.24
#> 130  618.596992   59623.18
#> 131  623.674624   59697.86
#> 132  628.797568   58954.22
#> 133  633.838272   63188.91
#> 134  638.875648   65002.90
#> 135  643.814336   69145.83
#> 136  648.777920   71580.73
#> 137  653.706496   74913.37
#> 138  658.598656   78510.85
#> 139  663.522688   82318.05
#> 140  668.365888   82551.84
#> 141  673.210048   85705.81
#> 142  678.053376   85455.64
#> 143  682.936576   81470.25
#> 144  687.832768   85282.35
#> 145  692.684480   89828.83
#> 146  697.507008   93287.83
#> 147  702.301888   94535.90
#> 148  707.104960   94970.70
#> 149  711.891840   96463.55
#> 150  716.724160   94340.63
#> 151  721.515456   95439.87
#> 152  726.289216  106277.11
#> 153  730.911232  122559.37
#> 154  735.468352  145943.00
#> 155  739.937792  166197.92
#> 156  744.378176  189759.50
#> 157  748.808960  209995.04
#> 158  753.194752  215009.91
#> 159  757.592192  233263.87
#> 160  761.971136  236090.90
#> 161  766.339328  243688.71
#> 162  770.760960  216306.01
#> 163  775.216960  185552.40
#> 164  779.771072  162308.84
#> 165  784.355200  146846.64
#> 166  789.002368  133857.68
#> 167  793.684864  124450.39
#> 168  798.392832  116767.95
#> 169  803.135104  113442.87
#> 170  807.886080  108957.80
#> 171  812.641280  106527.07
#> 172  817.379584  107631.19
#> 173  822.110016  116366.56
#> 174  826.841664  113201.29
#> 175  831.572224  118963.41
#> 176  836.268736  127917.80
#> 177  840.926272  131353.95
#> 178  845.580544  135252.47
#> 179  850.164928  146298.00
#> 180  854.635008  192447.34
#> 181  858.999872  260827.57
#> 182  863.264192  321581.14
#> 183  867.539520  370770.73
#> 184  871.813696  349936.67
#> 185  876.108032  294436.50
#> 186  880.512832  231991.97
#> 187  884.978112  186280.12
#> 188  889.540480  157089.21
#> 189  894.204928  132505.73
#> 190  898.957440  116660.99
#> 191  903.750208  107788.84
#> 192  908.588864  101021.13
#> 193  913.446016   97800.81
#> 194  918.300992  102978.61
#> 195  922.952256  127745.26
#> 196  927.442304  188394.66
#> 197  931.767872  260117.64
#> 198  936.046976  340604.32
#> 199  940.290624  358533.60
#> 200  944.543232  336077.03
#> 201  948.887040  277242.27
#> 202  953.335232  217200.67
#> 203  957.847488  168253.29
#> 204  962.507136  140789.81
#> 205  967.159360  128257.30
#> 206  971.881728  115171.61
#> 207  976.634752  103188.18
#> 208  981.488768   98175.15
#> 209  986.302464  102043.96
#> 210  991.026112  114109.96
#> 211  995.661824  128435.48
#> 212 1000.284672  127953.18
#> 213 1004.868416  137429.82
#> 214 1009.409856  153326.05
#> 215 1013.814400  206864.54
#> 216 1018.555520  362867.28
#> 217 1022.979264  575597.65
#> 218 1027.141888  684520.83
#> 219 1031.302976  730435.97
#> 220 1035.462464  624362.14
#> 221 1039.675392  517777.14
#> 222 1043.973120  407505.06
#> 223 1048.328704  290323.34
#> 224 1052.774528  234989.20
#> 225 1057.226368  225201.75
#> 226 1061.620672  240722.26
#> 227 1065.918016  332313.53
#> 228 1070.108608  517978.64
#> 229 1074.507712  732430.33
#> 230 1078.885952  926731.72
#> 231 1082.994752  994616.22
#> 232 1087.146432  908199.75
#> 233 1091.321216  739446.05
#> 234 1095.502144  575088.93
#> 235 1099.762688  445203.15
#> 236 1104.090624  350595.03
#> 237 1108.465536  271812.61
#> 238 1112.979712  203197.37
#> 239 1117.568960  163833.66
#> 240 1122.315776  127398.95
#> 241 1127.153536  106236.16
#> 242 1132.057856   95152.26
#> 243 1136.911744   90187.64
#> 244 1141.836544   87571.53
#> 245 1146.756672   84172.46
#> 246 1151.741952   81111.49
#> 247 1156.719232   80040.83
#> 248 1161.544384  107370.93
#> 249 1166.330560  224611.17
#> 250 1170.687232  367983.90
#> 251 1175.004352  428962.73
#> 252 1179.352192  375034.82
#> 253 1183.737024  310060.06
#> 254 1188.134016  354974.78
#> 255 1192.447232  418547.53
#> 256 1196.710912  447069.12
#> 257 1200.977280  472525.73
#> 258 1205.230912  481841.35
#> 259 1209.495296  493100.36
#> 260 1213.779328  406758.15
#> 261 1218.136640  311575.61
#> 262 1222.607872  263433.76
#> 263 1227.109056  226467.08
#> 264 1231.640448  204146.08
#> 265 1236.210240  169384.06
#> 266 1240.855488  137598.58
#> 267 1245.618688  116809.26
#> 268 1250.404672  103410.60
#> 269 1255.266048   93191.13
#> 270 1260.223680   83620.31
#> 271 1265.240000   78707.50
#> 272 1270.346880   71445.97
#> 273 1275.424896   70748.66
#> 274 1280.513600   68396.07
#> 275 1285.619648   66961.18
#> 276 1290.726592   67662.05
#> 277 1295.828800   67178.65
#> 278 1300.897792   66180.76
#> 279 1305.993984   63297.52
#> 280 1311.009856   72541.55
#> 281 1315.867584   85647.39
#> 282 1320.687488   91067.89
#> 283 1325.445120   99014.39
#> 284 1330.256512  104645.39
#> 285 1335.029760  104754.62
#> 286 1339.853504   92608.81
#> 287 1344.746304   88979.11
#> 288 1349.720064   81291.10
#> 289 1354.723264   75711.99
#> 290 1359.744192   76587.92
#> 291 1364.771712   75672.48
#> 292 1369.734528   77720.50
#> 293 1374.595584  100785.40
#> 294 1379.287744  137561.18
#> 295 1383.875200  186193.94
#> 296 1388.373632  188611.79
#> 297 1392.905088  216783.05
#> 298 1397.367872  205586.09
#> 299 1401.918592  170369.81
#> 300 1406.608256  141006.05
#> 301 1411.360768  118621.71
#> 302 1416.140544  108963.46
#> 303 1420.948864  104001.91
#> 304 1425.782528   93453.38
#> 305 1430.760896   82475.58
#> 306 1435.775232   77352.82
#> 307 1440.818112   72352.85
#> 308 1445.919040   69948.73
#> 309 1451.035328   70207.09
#> 310 1456.016832   83612.53
#> 311 1460.800704  106663.22
#> 312 1465.414528  175652.76
#> 313 1470.045632  351388.30
#> 314 1474.378112  504688.97
#> 315 1478.645056  537765.63
#> 316 1482.929152  477621.52
#> 317 1487.255808  371414.89
#> 318 1491.659328  269015.34
#> 319 1496.203776  211908.65
#> 320 1500.768896  161080.75
#> 321 1505.330880  166539.24
#> 322 1509.831552  186528.45
#> 323 1514.360576  182311.99
#> 324 1518.915648  161533.24
#> 325 1523.512768  143681.90
#> 326 1528.224064  123789.99
#> 327 1532.965696  105696.80
#> 328 1537.820352   96644.61
#> 329 1542.667200   87871.54
#> 330 1547.549056   84636.51
#> 331 1552.500672   78994.01
#> 332 1557.515072   75711.75
#> 333 1562.541376   81441.93
#> 334 1567.371328   98699.01
#> 335 1572.176960   98610.73
#> 336 1577.029824   92366.64
#> 337 1582.025984   83919.25
#> 338 1587.135552   73787.43
#> 339 1592.337600   65538.18
#> 340 1597.633600   59765.21
#> 341 1602.911680   58472.84
#> 342 1608.187520   56595.95
#> 343 1613.455424   56904.63
#> 344 1618.659200   57747.14
#> 345 1623.867328   57897.07
#> 346 1629.155200   53783.89
#> 347 1634.566400   50909.06
#> 348 1639.963520   49425.24
#> 349 1645.343168   53959.65
#> 350 1650.370176   82400.40
#> 351 1654.986240  119115.35
#> 352 1659.547456  131508.32
#> 353 1664.169728  115687.69
#> 354 1668.958592  101399.88
#> 355 1673.849152   88483.01
#> 356 1678.856704   76297.70
#> 357 1683.996672   64774.21
#> 358 1689.300800   55672.91
#> 359 1694.732864   50739.87
#> 360 1700.186752   47983.40
#> 361 1705.566528   54945.94
#> 362 1710.677184   66119.84
#> 363 1715.496768   81601.30
#> 364 1720.313088   77187.06
#> 365 1725.295616   72575.39
#> 366 1730.386112   68145.93
#> 367 1735.562304   63608.05
#> 368 1740.677760   63831.35
#> 369 1745.770240   65516.85
#> 370 1750.950208   60456.55
#> 371 1756.238080   54647.93
#> 372 1761.640384   50354.49
#> 373 1767.126272   45482.65
#> 374 1772.642496   43529.79
#> 375 1778.178368   44423.97
#> 376 1783.638592   46266.42
#> 377 1789.050368   47063.19
#> 378 1794.434240   45803.59
#> 379 1799.864256   46843.22
#> 380 1805.254080   46886.41
#> 381 1810.687872   47904.44
#> 382 1816.075968   47676.05
#> 383 1821.482176   47787.85
#> 384 1826.913792   47308.19
#> 385 1832.356288   48321.79
#> 386 1837.741760   49130.56
#> 387 1843.111680   47870.43
#> 388 1848.543680   47362.65
#> 389 1853.987904   47217.77
#> 390 1859.427456   46168.20
#> 391 1864.717312   55505.76
#> 392 1869.597568   84691.69
#> 393 1874.375168  131942.18
#> 394 1878.782208  191554.91
#> 395 1883.157056  191530.86
#> 396 1887.586112  165750.24
#> 397 1892.171840  140090.68
#> 398 1896.826048  121327.45
#> 399 1901.565440  112687.77
#> 400 1906.310464  108868.94
#> 401 1911.069568  110996.84
#> 402 1915.689600  129269.27
#> 403 1920.264960  149105.14
#> 404 1924.795904  154312.05
#> 405 1929.353792  135474.89
#> 406 1934.029056  117523.45
#> 407 1938.803136  101498.90
#> 408 1943.673984   88664.15
#> 409 1948.706112   76074.56
#> 410 1953.832640   64211.80
#> 411 1959.140800   54668.29
#> 412 1964.634688   48214.45
#> 413 1970.226880   43873.02
#> 414 1975.862976   42556.25
#> 415 1981.508992   40201.34
#> 416 1987.228992   39175.31
#> 417 1992.926720   38901.75
#> 418 1998.582656   38283.66
#> 419 2004.260544   38751.70
#> 420 2009.914368   39357.24
#> 421 2015.636800   39155.16
#> 422 2021.322560   39230.77
#> 423 2026.950464   38769.78
#> 424 2032.610496   39423.34
#> 425 2038.264512   39125.75
#> 426 2043.832320   42355.28
#> 427 2049.280064   46221.68
#> 428 2054.576000   49845.77
#> 429 2059.864000   51486.00
#> 430 2065.248000   48443.00
#> 431 2070.706048   48065.68
#> 432 2076.212224   46099.71
#> 433 2081.686016   45012.36
#> 434 2087.233984   44147.47
#> 435 2092.769984   43788.97
#> 436 2098.310080   44142.87
#> 437 2103.825792   43356.66
#> 438 2109.339968   43987.05
#> 439 2114.886400   44189.59
#> 440 2120.333760   46825.30
#> 441 2125.695360   48837.59
#> 442 2131.035648   52126.59
#> 443 2136.277568   53183.08
#> 444 2141.561344   51745.77
#> 445 2146.903552   50327.97
#> 446 2152.301376   48255.12
#> 447 2157.751488   46918.54
#> 448 2163.215296   46523.98
#> 449 2168.639488   45929.40
#> 450 2174.053440   48584.29
#> 451 2179.425216   48169.87
#> 452 2184.822848   47301.06
#> 453 2190.287104   45995.11
#> 454 2195.719040   46384.20
#> 455 2201.201024   45786.12
#> 456 2206.683200   46334.28
#> 457 2212.049088   50354.61
#> 458 2217.307136   52662.05
#> 459 2222.575040   53381.38
#> 460 2227.824704   54078.06
#> 461 2233.094784   54280.00
#> 462 2238.460928   51904.37
#> 463 2243.866816   49356.83
#> 464 2249.275264   47456.07
#> 465 2254.720640   47382.99
#> 466 2260.198528   46674.77
#> 467 2265.669184   45116.29
#> 468 2271.142656   45455.48
#> 469 2276.634560   44528.95
#> 470 2282.128256   44063.62
#> 471 2287.649856   44336.51
#> 472 2293.158528   43548.20
#> 473 2298.620416   42941.77
#> 474 2304.096576   44148.26
#> 475 2309.589760   46215.32
#> 476 2315.044224   43066.70
#> 477 2320.540352   43565.98
#> 478 2326.008512   45019.16
#> 479 2331.449408   45087.83
#> 480 2336.876288   45787.54
#> 481 2342.302016   47123.65
#> 482 2347.712192   48574.73
#> 483 2353.092416   46301.68
#> 484 2358.568320   44779.68
#> 485 2364.071872   43835.57
#> 486 2369.571968   45193.31
#> 487 2375.013888   44603.26
#> 488 2380.501888   45317.58
#> 489 2385.961856   45716.24
#> 490 2391.422080   45529.95
#> 491 2396.877824   43862.77
#> 492 2402.339776   45434.56
#> 493 2407.779648   46418.38
#> 494 2413.196032   45907.84
#> 495 2418.561536   47039.41
#> 496 2423.915648   46996.97
#> 497 2429.304128   48438.62
#> 498 2434.619456   50471.28
#> 499 2439.865536   53053.92
#> 500 2445.149376   53428.58
#> 501 2450.421312   50918.42
#> 502 2455.749312   50701.49
#> 503 2461.079296   50963.57
#> 504 2466.371456   52943.24
#> 505 2471.639296   51654.95
#> 506 2476.883136   55734.94
#> 507 2482.033344   56022.19
#> 508 2487.211072   55835.08
#> 509 2492.399040   57224.72
#> 510 2497.595264   56443.04
#> 511 2502.803072   56738.19
#> 512 2507.952960   59225.25
#> 513 2513.044992   62127.79
#> 514 2518.187072   60047.72
#> 515 2523.372800   59330.19
#> 516 2528.514752   59693.52
#> 517 2533.668992   59145.99
#> 518 2538.825024   59779.14
#> 519 2543.886592   62569.35
#> 520 2548.932608   68022.22
#> 521 2553.964992   68668.81
#> 522 2558.937920   69814.86
#> 523 2563.990208   67422.04
#> 524 2569.020672   65641.66
#> 525 2574.084928   65275.09
#> 526 2579.115968   65053.55
#> 527 2584.106880   68476.02
#> 528 2589.103488   68299.03
#> 529 2594.064576   71033.08
#> 530 2599.012416   76084.63
#> 531 2603.875392   76526.03
#> 532 2608.764032   80566.55
#> 533 2613.661632   76500.62
#> 534 2618.561920   75960.89
#> 535 2623.469248   72094.24
#> 536 2628.396416   71223.72
#> 537 2633.335680   71694.06
#> 538 2638.213248   74341.34
#> 539 2643.023808   86676.65
#> 540 2647.763968   88357.43
#> 541 2652.574336   85095.50
#> 542 2657.398400   81040.42
#> 543 2662.233088   79972.96
#> 544 2667.100608   79071.06
#> 545 2671.931968   82452.68
#> 546 2676.739904   84118.81
#> 547 2681.522368   84053.92
#> 548 2686.365376   85188.22
#> 549 2691.142336   91551.69
#> 550 2695.832576  107281.06
#> 
#> [[2]]
#>           rtime  intensity
#> 1      0.295936   27936.73
#> 2      6.265088   28351.22
#> 3     12.173184   29381.87
#> 4     18.077248   29711.06
#> 5     23.986944   29711.95
#> 6     29.897024   29454.03
#> 7     35.809024   29301.50
#> 8     41.716864   29631.17
#> 9     47.644800   29514.92
#> 10    53.568576   28538.80
#> 11    59.480768   28271.78
#> 12    65.378048   29137.62
#> 13    71.270784   28114.20
#> 14    77.190720   29192.84
#> 15    83.084544   28230.08
#> 16    89.026496   28761.95
#> 17    94.964416   27862.27
#> 18   100.898368   28157.81
#> 19   106.816704   28974.50
#> 20   112.758336   29541.64
#> 21   118.660416   29196.25
#> 22   124.559104   30630.91
#> 23   130.017152  138945.76
#> 24   135.002048  513478.13
#> 25   139.123200  788318.15
#> 26   143.257536  880791.98
#> 27   147.413184  758189.81
#> 28   151.557056  763188.43
#> 29   155.725184  872934.68
#> 30   159.883264  935753.59
#> 31   164.003072 1020918.05
#> 32   168.158016  995544.29
#> 33   172.320512  957987.74
#> 34   176.462720  929367.44
#> 35   180.645376  775598.00
#> 36   184.861440  634835.05
#> 37   189.066112  513530.13
#> 38   193.340160  419279.70
#> 39   197.643392  332282.78
#> 40   201.955264  272621.30
#> 41   206.318720  233436.62
#> 42   210.708608  218137.84
#> 43   215.094016  206129.26
#> 44   219.502976  201851.58
#> 45   223.896192  189846.14
#> 46   228.312000  187942.36
#> 47   232.732672  189863.29
#> 48   237.129344  195877.85
#> 49   241.532544  204997.57
#> 50   245.894144  204799.83
#> 51   250.279424  206059.24
#> 52   254.680192  206676.51
#> 53   259.074944  207005.12
#> 54   263.496192  214932.39
#> 55   267.885056  212859.36
#> 56   272.309248  218636.95
#> 57   276.732160  211730.11
#> 58   281.194240  208788.97
#> 59   285.677120  202164.80
#> 60   290.137344  209323.79
#> 61   294.553920  283886.15
#> 62   298.830336  410078.29
#> 63   303.082048  522911.56
#> 64   307.326464  585068.68
#> 65   311.548928  575152.93
#> 66   315.800832  568997.90
#> 67   320.054784  577691.78
#> 68   324.285312  577152.25
#> 69   328.527680  603744.15
#> 70   332.775744  624510.74
#> 71   337.022592  594080.39
#> 72   341.287872  548791.77
#> 73   345.562368  483646.44
#> 74   349.849728  408175.54
#> 75   354.170496  353800.94
#> 76   358.530368  312381.02
#> 77   362.944128  272858.99
#> 78   367.382336  229233.73
#> 79   371.865792  194928.11
#> 80   376.401856  176874.33
#> 81   380.946944  157578.65
#> 82   385.545728  150653.50
#> 83   390.136384  143634.55
#> 84   394.723776  146274.99
#> 85   399.295552  160819.63
#> 86   403.852864  159559.96
#> 87   408.412864  155803.00
#> 88   412.974144  145538.23
#> 89   417.594944  137071.68
#> 90   422.259968  122502.45
#> 91   427.035648  113122.31
#> 92   431.796480  102863.74
#> 93   436.570368  102055.11
#> 94   441.344960   98038.96
#> 95   446.134336   99290.71
#> 96   450.910656  100097.87
#> 97   455.667712  104224.68
#> 98   460.435648  103636.60
#> 99   465.212544  103638.15
#> 100  469.967744  113736.44
#> 101  474.657472  130789.27
#> 102  479.280384  145908.74
#> 103  483.890688  155540.63
#> 104  488.507264  156060.06
#> 105  493.123264  158474.25
#> 106  497.711872  155082.98
#> 107  502.331264  161975.12
#> 108  506.920704  159211.47
#> 109  511.553472  153307.66
#> 110  516.145600  151175.85
#> 111  520.803200  148456.64
#> 112  525.449728  135949.37
#> 113  530.179648  116974.63
#> 114  534.933248  105498.96
#> 115  539.811392   93564.65
#> 116  544.708224   82984.49
#> 117  549.650240   79279.09
#> 118  554.629120   75575.85
#> 119  559.610304   73947.36
#> 120  564.601984   70633.60
#> 121  569.574528   69625.82
#> 122  574.592192   67793.22
#> 123  579.609920   64959.01
#> 124  584.665408   62911.90
#> 125  589.754816   62378.98
#> 126  594.825344   60105.97
#> 127  599.963712   58939.57
#> 128  605.081856   56906.13
#> 129  610.207424   57128.07
#> 130  615.353216   57530.13
#> 131  620.439744   60181.02
#> 132  625.463936   61404.62
#> 133  630.497152   63781.37
#> 134  635.463296   68945.49
#> 135  640.379072   74689.45
#> 136  645.246144   80908.42
#> 137  650.102272   85916.38
#> 138  654.949184   91131.60
#> 139  659.768384   92130.12
#> 140  664.600768   91640.19
#> 141  669.389120   93506.86
#> 142  674.174784   96751.73
#> 143  678.957952   98135.04
#> 144  683.750336   95976.14
#> 145  688.566656   94381.09
#> 146  693.408896   97569.00
#> 147  698.169024  105757.57
#> 148  702.921792  111751.27
#> 149  707.617088  114047.59
#> 150  712.341952  116689.02
#> 151  717.024000  120571.06
#> 152  721.731968  116649.23
#> 153  726.410880  116202.74
#> 154  731.078528  132482.95
#> 155  735.597312  156067.41
#> 156  740.018816  202479.78
#> 157  744.378176  246472.80
#> 158  748.695232  293900.88
#> 159  752.994496  333249.42
#> 160  757.241728  368133.27
#> 161  761.507712  378730.35
#> 162  765.761280  404219.16
#> 163  769.984384  420223.47
#> 164  774.243200  360663.83
#> 165  778.536768  310115.18
#> 166  782.881792  253264.58
#> 167  787.299904  214380.53
#> 168  791.793344  181146.37
#> 169  796.360256  155949.99
#> 170  801.019712  135057.79
#> 171  805.737664  120529.17
#> 172  810.479360  112563.13
#> 173  815.242368  114491.35
#> 174  819.913280  134982.39
#> 175  824.513344  161334.82
#> 176  829.076800  172747.45
#> 177  833.628672  193846.09
#> 178  838.160768  194711.66
#> 179  842.653632  202087.75
#> 180  847.217664  198641.57
#> 181  851.742336  175720.97
#> 182  856.289728  177848.08
#> 183  860.693056  240879.33
#> 184  864.991360  323336.17
#> 185  869.238848  404262.17
#> 186  873.432768  423038.62
#> 187  877.676736  384721.58
#> 188  881.958208  338158.38
#> 189  886.260864  278591.55
#> 190  890.654144  239236.44
#> 191  895.118400  205984.99
#> 192  899.662848  182154.34
#> 193  904.265280  153252.64
#> 194  908.932544  136816.51
#> 195  913.711424  117725.62
#> 196  918.514880  106677.16
#> 197  923.292864  109354.18
#> 198  927.940544  140775.28
#> 199  932.360896  215872.98
#> 200  936.847872  364381.34
#> 201  941.056320  480966.97
#> 202  945.269632  485211.95
#> 203  949.463616  442759.05
#> 204  953.705728  379008.17
#> 205  958.050752  286282.37
#> 206  962.511680  213204.37
#> 207  967.080512  165449.42
#> 208  971.721664  136101.82
#> 209  976.430528  117018.76
#> 210  981.172224  105536.10
#> 211  985.950336   96633.10
#> 212  990.636800  115440.14
#> 213  995.180928  152533.89
#> 214  999.602816  193973.66
#> 215 1004.017472  204822.26
#> 216 1008.414656  204925.50
#> 217 1012.812352  240062.48
#> 218 1017.115840  327422.66
#> 219 1021.757120  628255.72
#> 220 1026.133888  840158.48
#> 221 1030.233600 1049849.46
#> 222 1034.336192 1035307.92
#> 223 1038.446912  936751.97
#> 224 1042.608832  790018.34
#> 225 1046.770816  635495.54
#> 226 1050.999040  473449.76
#> 227 1055.290880  355552.39
#> 228 1059.595904  329462.90
#> 229 1063.963520  336067.21
#> 230 1068.249216  408627.45
#> 231 1072.688320  550551.54
#> 232 1077.078528  774787.14
#> 233 1081.423808  996729.80
#> 234 1085.566848 1096781.15
#> 235 1089.697856 1027688.01
#> 236 1093.819136  882520.21
#> 237 1097.993088  732437.56
#> 238 1102.207808  570460.08
#> 239 1106.451136  399868.73
#> 240 1110.836928  286322.05
#> 241 1115.360448  206249.42
#> 242 1120.008512  155807.52
#> 243 1124.772160  121723.20
#> 244 1129.645120  103726.32
#> 245 1134.519296   91294.44
#> 246 1139.443328   90671.60
#> 247 1144.332480   85588.60
#> 248 1149.258688   83014.07
#> 249 1154.241280   78698.08
#> 250 1159.265984   75436.65
#> 251 1164.237120   84922.68
#> 252 1169.120640  191434.17
#> 253 1173.759104  379090.11
#> 254 1178.078848  505875.92
#> 255 1182.391168  508805.05
#> 256 1186.723200  412524.25
#> 257 1191.090112  345766.33
#> 258 1195.451072  386343.57
#> 259 1199.765376  413837.50
#> 260 1204.041216  432647.75
#> 261 1208.325312  416614.33
#> 262 1212.597760  419735.65
#> 263 1216.908352  375345.84
#> 264 1221.284928  296643.81
#> 265 1225.739328  236811.09
#> 266 1230.299136  203032.31
#> 267 1234.900608  177366.39
#> 268 1239.497664  154327.18
#> 269 1244.123008  129866.22
#> 270 1248.801984  117661.90
#> 271 1253.501824  108950.36
#> 272 1258.255424   96771.82
#> 273 1263.120320   86007.70
#> 274 1268.087744   79270.22
#> 275 1273.103232   72896.96
#> 276 1278.190912   72083.68
#> 277 1283.233984   70166.08
#> 278 1288.280896   68467.00
#> 279 1293.444224   66174.21
#> 280 1298.555072   68718.78
#> 281 1303.618816   69329.30
#> 282 1308.670976   68520.53
#> 283 1313.720832   66837.30
#> 284 1318.713984   73451.92
#> 285 1323.522624   91258.73
#> 286 1328.282880  115495.49
#> 287 1332.905088  133541.95
#> 288 1337.552256  146754.88
#> 289 1342.146688  140715.54
#> 290 1346.854016  120917.16
#> 291 1351.653760  105294.02
#> 292 1356.580032   90066.06
#> 293 1361.550144   84122.52
#> 294 1366.533248   78642.93
#> 295 1371.508352   77289.87
#> 296 1376.453312   88563.20
#> 297 1381.255680  118734.00
#> 298 1385.842496  182565.53
#> 299 1390.356928  231333.14
#> 300 1394.852160  221144.34
#> 301 1399.388288  202880.11
#> 302 1404.000576  170165.54
#> 303 1408.722432  133330.94
#> 304 1413.545984  107615.76
#> 305 1418.407744  103741.12
#> 306 1423.258432   93641.36
#> 307 1428.198400   85046.75
#> 308 1433.166464   79206.94
#> 309 1438.265728   71020.22
#> 310 1443.371136   68468.75
#> 311 1448.528896   63688.22
#> 312 1453.815040   59951.72
#> 313 1458.922496   73097.87
#> 314 1463.798016   96764.70
#> 315 1468.531200  107512.73
#> 316 1473.130432  167329.03
#> 317 1477.788032  298002.55
#> 318 1482.117184  391966.58
#> 319 1486.392704  373620.55
#> 320 1490.731136  306389.53
#> 321 1495.183616  226268.37
#> 322 1499.792192  164906.63
#> 323 1504.488576  138802.65
#> 324 1509.193600  132932.12
#> 325 1513.836928  144493.41
#> 326 1518.414400  155065.58
#> 327 1522.978240  155342.33
#> 328 1527.602880  136952.42
#> 329 1532.335936  120728.08
#> 330 1537.085888  109866.55
#> 331 1541.856960  105944.37
#> 332 1546.613824  103767.45
#> 333 1551.407808   99211.94
#> 334 1556.236160   90877.68
#> 335 1561.162432   87033.31
#> 336 1566.153920   78112.61
#> 337 1571.212416   76176.34
#> 338 1576.305344   74061.04
#> 339 1581.416512   71979.89
#> 340 1586.559424   68079.91
#> 341 1591.806592   60591.60
#> 342 1597.121792   56246.35
#> 343 1602.472064   53558.73
#> 344 1607.883840   52095.07
#> 345 1613.276032   52725.45
#> 346 1618.609920   53811.75
#> 347 1623.861760   57008.19
#> 348 1629.063616   58496.74
#> 349 1634.329472   56579.55
#> 350 1639.683584   52868.49
#> 351 1645.165696   50925.78
#> 352 1650.591616   51465.95
#> 353 1655.797184   66626.15
#> 354 1660.611200  101550.16
#> 355 1665.240832  114335.81
#> 356 1669.915904  103551.09
#> 357 1674.735104   92360.31
#> 358 1679.736896   80957.05
#> 359 1684.768000   73655.76
#> 360 1689.940928   65741.86
#> 361 1695.222720   57909.11
#> 362 1700.573248   54213.80
#> 363 1706.015424   50547.87
#> 364 1711.407104   51763.80
#> 365 1716.516864   64749.34
#> 366 1721.313280   91440.00
#> 367 1726.021888   99144.75
#> 368 1730.842752   84058.27
#> 369 1735.823296   77825.92
#> 370 1740.899392   72509.18
#> 371 1746.043200   66904.68
#> 372 1751.146624   66427.32
#> 373 1756.328448   59539.40
#> 374 1761.658816   55327.33
#> 375 1767.080896   49249.28
#> 376 1772.552896   46306.64
#> 377 1778.064640   47599.86
#> 378 1783.512448   47797.18
#> 379 1788.916544   48639.07
#> 380 1794.316480   47408.78
#> 381 1799.750336   46815.91
#> 382 1805.256384   45949.39
#> 383 1810.730304   45047.03
#> 384 1816.230656   44968.06
#> 385 1821.686208   44912.35
#> 386 1827.190464   45748.04
#> 387 1832.586304   47787.86
#> 388 1837.918144   50926.56
#> 389 1843.164032   57311.51
#> 390 1848.297856   61786.34
#> 391 1853.467712   57985.70
#> 392 1858.781824   52323.39
#> 393 1864.178112   50717.81
#> 394 1869.548416   53554.99
#> 395 1874.513600   79245.85
#> 396 1879.323584  140169.47
#> 397 1883.704704  214089.24
#> 398 1887.996352  263664.75
#> 399 1892.333248  243353.21
#> 400 1896.733248  198718.71
#> 401 1901.208576  165686.94
#> 402 1905.817280  149604.62
#> 403 1910.418624  134539.29
#> 404 1915.067648  124138.85
#> 405 1919.650496  139920.66
#> 406 1924.229824  158719.23
#> 407 1928.667648  191604.29
#> 408 1933.105664  198709.98
#> 409 1937.565248  174302.13
#> 410 1942.151424  142606.49
#> 411 1946.841984  121597.15
#> 412 1951.605120  100817.45
#> 413 1956.496128   90551.06
#> 414 1961.538880   73933.04
#> 415 1966.749248   62556.77
#> 416 1972.146816   53492.52
#> 417 1977.699008   46310.90
#> 418 1983.294912   43387.82
#> 419 1988.971264   40104.20
#> 420 1994.716288   39365.98
#> 421 2000.444928   40299.35
#> 422 2006.122688   39612.00
#> 423 2011.815232   38982.77
#> 424 2017.563008   39346.20
#> 425 2023.212736   39566.92
#> 426 2028.936128   38488.17
#> 427 2034.718720   37219.41
#> 428 2040.468800   37624.86
#> 429 2046.178624   38704.88
#> 430 2051.793472   41224.08
#> 431 2057.240448   45755.99
#> 432 2062.588288   51196.98
#> 433 2067.910464   48672.94
#> 434 2073.346240   48523.65
#> 435 2078.856576   47952.71
#> 436 2084.388416   45570.48
#> 437 2089.904512   45027.58
#> 438 2095.426240   43356.28
#> 439 2101.012224   42674.64
#> 440 2106.630208   43247.34
#> 441 2112.188352   43324.64
#> 442 2117.736256   43517.69
#> 443 2123.256128   44689.22
#> 444 2128.736448   47478.53
#> 445 2134.077952   49124.73
#> 446 2139.367872   51681.10
#> 447 2144.661888   53269.32
#> 448 2150.012480   50203.99
#> 449 2155.443840   48389.34
#> 450 2160.927680   47342.25
#> 451 2166.397760   47471.81
#> 452 2171.808064   48417.80
#> 453 2177.189568   49609.50
#> 454 2182.589696   50139.39
#> 455 2187.985536   50797.31
#> 456 2193.347328   50318.66
#> 457 2198.809792   47753.34
#> 458 2204.347520   45655.61
#> 459 2209.850304   44401.10
#> 460 2215.285760   46545.21
#> 461 2220.641728   51370.17
#> 462 2225.951232   53621.21
#> 463 2231.153024   55406.83
#> 464 2236.455424   50950.67
#> 465 2241.855168   51209.66
#> 466 2247.291456   48903.37
#> 467 2252.733248   47017.99
#> 468 2258.207168   48180.62
#> 469 2263.692800   43957.25
#> 470 2269.215360   44567.37
#> 471 2274.762880   42433.58
#> 472 2280.355200   42652.53
#> 473 2285.967104   41236.02
#> 474 2291.586944   41629.29
#> 475 2297.106624   41914.18
#> 476 2302.630656   43310.04
#> 477 2308.204672   40633.77
#> 478 2313.842816   40531.64
#> 479 2319.403008   43147.71
#> 480 2324.906496   43305.00
#> 481 2330.392640   43412.19
#> 482 2335.930432   43673.97
#> 483 2341.432448   44041.61
#> 484 2346.942720   46136.50
#> 485 2352.390208   46278.02
#> 486 2357.910592   44326.04
#> 487 2363.418432   44149.65
#> 488 2368.956352   43805.82
#> 489 2374.468096   43130.99
#> 490 2380.024320   44018.71
#> 491 2385.548224   43267.72
#> 492 2391.076160   42880.68
#> 493 2396.600512   44599.10
#> 494 2402.132032   43216.54
#> 495 2407.676480   43602.19
#> 496 2413.182080   43878.68
#> 497 2418.631744   44894.61
#> 498 2424.083968   46301.91
#> 499 2429.525184   45654.88
#> 500 2434.899904   47320.89
#> 501 2440.245824   49770.62
#> 502 2445.551552   52604.40
#> 503 2450.877824   49034.04
#> 504 2456.309632   46416.79
#> 505 2461.764032   47512.65
#> 506 2467.168640   46554.78
#> 507 2472.545664   47613.23
#> 508 2477.869696   51552.16
#> 509 2483.107584   54838.22
#> 510 2488.223296   58276.71
#> 511 2493.397504   56198.16
#> 512 2498.729536   55359.37
#> 513 2503.989248   52725.90
#> 514 2509.287360   51949.03
#> 515 2514.597184   53102.21
#> 516 2519.861440   53193.68
#> 517 2525.149120   52034.82
#> 518 2530.483136   51799.00
#> 519 2535.789312   51939.92
#> 520 2541.061376   53627.27
#> 521 2546.293376   55904.34
#> 522 2551.453056   56900.76
#> 523 2556.522368   65069.14
#> 524 2561.572992   69589.60
#> 525 2566.600832   65881.15
#> 526 2571.702784   65690.41
#> 527 2576.832960   61008.51
#> 528 2581.996864   59422.78
#> 529 2587.124992   62083.08
#> 530 2592.178688   61014.81
#> 531 2597.271872   62440.96
#> 532 2602.422784   61759.17
#> 533 2607.516480   60353.11
#> 534 2612.624384   59896.78
#> 535 2617.735040   60363.93
#> 536 2622.842432   59509.20
#> 537 2627.934464   57876.66
#> 538 2633.004608   58732.55
#> 539 2638.052480   60029.58
#> 540 2643.122560   64222.88
#> 541 2648.050880   72558.74
#> 542 2652.992256   71626.12
#> 543 2657.917824   70669.70
#> 544 2662.877952   70939.49
#> 545 2667.840640   69815.28
#> 546 2672.825152   68065.23
#> 547 2677.872256   65278.46
#> 548 2682.848576   66309.73
#> 549 2687.833344   67184.52
#> 550 2692.821952   69864.77
#> 551 2697.736192   76635.44
#> 
#> [[3]]
#>           rtime  intensity
#> 1      3.660864   32484.68
#> 2      9.446720   32471.46
#> 3     15.248960   32050.10
#> 4     21.074688   32182.25
#> 5     26.860544   32716.69
#> 6     32.680704   32643.15
#> 7     38.440704   32668.45
#> 8     44.240576   33246.91
#> 9     50.034432   32653.30
#> 10    55.838720   32980.50
#> 11    61.652736   32641.17
#> 12    67.468672   32753.93
#> 13    73.245248   32715.86
#> 14    79.014912   33617.10
#> 15    84.756288   33281.67
#> 16    90.554176   34696.68
#> 17    96.286208   33589.56
#> 18   102.114176   33359.54
#> 19   107.900160   33019.07
#> 20   113.677952   32898.06
#> 21   119.435904   34156.36
#> 22   125.240384   32149.24
#> 23   130.747072   88240.78
#> 24   136.229824  598246.13
#> 25   140.338112  959219.15
#> 26   144.466048  970202.67
#> 27   148.606528  945865.50
#> 28   152.761280  907751.45
#> 29   156.905856  982662.44
#> 30   161.043584 1076208.89
#> 31   165.170112 1149714.05
#> 32   169.286976 1108058.26
#> 33   173.434880  903304.00
#> 34   177.628608  767425.60
#> 35   181.815296  649102.17
#> 36   186.054848  542991.37
#> 37   190.326336  411030.63
#> 38   194.583360  329944.64
#> 39   198.962432  262009.98
#> 40   203.348672  216712.84
#> 41   207.761728  189310.19
#> 42   212.181824  169723.66
#> 43   216.594816  169167.41
#> 44   221.030080  168139.35
#> 45   225.437888  165217.70
#> 46   229.841024  173444.33
#> 47   234.225408  182756.85
#> 48   238.587776  198814.26
#> 49   242.925696  212150.36
#> 50   247.248384  233687.16
#> 51   251.592832  234719.24
#> 52   255.932416  245389.29
#> 53   260.250688  256354.29
#> 54   264.619456  249281.97
#> 55   268.978048  264334.11
#> 56   273.319296  271894.54
#> 57   277.732352  266692.50
#> 58   282.077056  278683.33
#> 59   286.460096  268775.49
#> 60   290.859904  257582.35
#> 61   295.258368  241389.41
#> 62   299.699712  214715.59
#> 63   304.119680  192292.27
#> 64   308.560640  178171.74
#> 65   313.003904  178859.32
#> 66   317.409408  178042.02
#> 67   321.890112  174421.90
#> 68   326.331392  167513.49
#> 69   330.819584  156254.70
#> 70   335.306368  155004.15
#> 71   339.796864  151895.73
#> 72   344.278784  148311.27
#> 73   348.753920  150471.47
#> 74   353.326848  131845.01
#> 75   357.920256  115070.40
#> 76   362.663232  103642.28
#> 77   367.432832   92956.31
#> 78   372.275200   86439.33
#> 79   377.107840   83298.37
#> 80   381.977920   79781.71
#> 81   386.822080   80650.32
#> 82   391.650944   84541.54
#> 83   396.512640   80691.26
#> 84   401.403648   77377.84
#> 85   406.318464   74808.45
#> 86   411.224576   77371.17
#> 87   416.130688   76118.14
#> 88   420.959552   82525.68
#> 89   425.791744   83532.88
#> 90   430.630528   80543.66
#> 91   435.499136   81279.99
#> 92   440.393792   78143.85
#> 93   445.332160   73522.35
#> 94   450.300096   71728.41
#> 95   455.283328   67768.64
#> 96   460.276288   71430.10
#> 97   465.190208   75960.05
#> 98   470.131776   73815.94
#> 99   475.083456   73732.51
#> 100  480.061760   69288.79
#> 101  485.118528   66089.76
#> 102  490.162560   66049.16
#> 103  495.219328   64876.70
#> 104  500.303808   62862.67
#> 105  505.401600   61554.25
#> 106  510.557696   58659.53
#> 107  515.669376   58063.37
#> 108  520.791360   58475.05
#> 109  525.947712   57577.69
#> 110  531.069248   55089.80
#> 111  536.209472   55209.45
#> 112  541.347328   55017.10
#> 113  546.509376   55461.07
#> 114  551.687168   55886.69
#> 115  556.809280   56168.15
#> 116  561.956992   54693.28
#> 117  567.187264   53001.13
#> 118  572.443328   50088.65
#> 119  577.704960   51202.92
#> 120  582.987008   50480.03
#> 121  588.220928   51039.99
#> 122  593.431168   49542.12
#> 123  598.675072   47808.24
#> 124  603.896896   48070.70
#> 125  609.169024   47456.54
#> 126  614.429120   47207.82
#> 127  619.662720   48095.25
#> 128  624.872576   49138.41
#> 129  630.114688   48194.57
#> 130  635.340672   46276.01
#> 131  640.584768   47966.26
#> 132  645.876544   47466.06
#> 133  651.134592   47016.36
#> 134  656.390272   46041.82
#> 135  661.646528   49877.27
#> 136  666.808192   53912.89
#> 137  671.909824   58246.46
#> 138  676.970048   63527.59
#> 139  681.913984   68097.24
#> 140  686.862080   70284.21
#> 141  691.770816   75594.48
#> 142  696.637632   78771.34
#> 143  701.504896   81311.41
#> 144  706.374016   80772.25
#> 145  711.275392   81691.29
#> 146  716.138240   78273.53
#> 147  721.025856   75423.54
#> 148  725.925568   77170.95
#> 149  730.839296   76683.39
#> 150  735.741632   77489.92
#> 151  740.652480   75802.07
#> 152  745.585728   76557.24
#> 153  750.476288   83143.73
#> 154  755.302528   87830.76
#> 155  759.987584  110767.99
#> 156  764.582016  145413.95
#> 157  768.994496  196943.27
#> 158  773.349760  258111.98
#> 159  777.671616  273897.45
#> 160  781.941376  314653.41
#> 161  786.229760  353583.15
#> 162  790.514368  372671.99
#> 163  794.761344  361542.68
#> 164  799.051264  291048.07
#> 165  803.463232  213108.18
#> 166  808.058304  166081.73
#> 167  812.745856  130866.41
#> 168  817.556224  110336.00
#> 169  822.445440   89760.50
#> 170  827.363136   83886.67
#> 171  832.249344   88733.50
#> 172  837.085888   92428.97
#> 173  841.854592   97707.81
#> 174  846.605888  111842.99
#> 175  851.240960  121695.76
#> 176  855.898048  125601.49
#> 177  860.524864  125927.08
#> 178  865.154752  136359.93
#> 179  869.730688  138884.59
#> 180  874.300800  147141.97
#> 181  878.764032  206419.20
#> 182  883.084416  269101.52
#> 183  887.369856  322089.15
#> 184  891.610752  346735.47
#> 185  895.908288  277748.56
#> 186  900.333376  203006.41
#> 187  904.943872  150198.36
#> 188  909.644096  133309.04
#> 189  914.287552  134102.49
#> 190  918.953984  135891.37
#> 191  923.554048  133087.41
#> 192  928.234496  127088.42
#> 193  932.980352  106125.45
#> 194  937.852224   95272.91
#> 195  942.657920  104581.93
#> 196  947.198080  165287.43
#> 197  951.511232  271934.03
#> 198  955.749760  400462.37
#> 199  959.968896  437364.18
#> 200  964.180032  411503.58
#> 201  968.489728  315089.96
#> 202  972.901696  230133.23
#> 203  977.495616  166704.06
#> 204  982.224512  126567.64
#> 205  987.058432  102357.68
#> 206  991.928000   92205.51
#> 207  996.829824   84387.95
#> 208 1001.733760   83882.55
#> 209 1006.631744   84715.61
#> 210 1011.315200  125351.21
#> 211 1015.834240  181253.48
#> 212 1020.227712  225278.23
#> 213 1024.651392  199619.66
#> 214 1029.140544  172502.48
#> 215 1033.599872  183961.85
#> 216 1038.190528  313306.51
#> 217 1042.589632  517354.65
#> 218 1046.754624  649039.00
#> 219 1050.908352  642906.73
#> 220 1055.068288  513388.56
#> 221 1059.346432  370027.61
#> 222 1063.764992  258009.26
#> 223 1068.321280  195017.76
#> 224 1072.827520  198455.89
#> 225 1077.388992  302925.54
#> 226 1081.634944  445762.97
#> 227 1085.813632  547834.56
#> 228 1090.224320  713018.43
#> 229 1094.607744  805284.02
#> 230 1098.722816  874574.66
#> 231 1102.872576  880005.83
#> 232 1107.036032  756972.12
#> 233 1111.214848  583379.00
#> 234 1115.502080  390696.24
#> 235 1119.940608  260941.56
#> 236 1124.545792  186675.38
#> 237 1129.338688  138410.58
#> 238 1134.156736  115214.37
#> 239 1138.983360  108501.58
#> 240 1143.748736  109056.54
#> 241 1148.560768   99572.22
#> 242 1153.438528   89884.74
#> 243 1158.392320   79525.36
#> 244 1163.410944   75424.87
#> 245 1168.382208   76859.77
#> 246 1173.361024   72620.08
#> 247 1178.392640   72225.27
#> 248 1183.436992   70649.76
#> 249 1188.456320   74688.68
#> 250 1193.414720  149600.48
#> 251 1198.098496  298012.28
#> 252 1202.370496  465245.88
#> 253 1206.658048  441532.62
#> 254 1210.988096  363246.20
#> 255 1215.399296  278133.67
#> 256 1220.069440  385992.00
#> 257 1224.336704  499204.95
#> 258 1228.614656  484177.72
#> 259 1232.891584  529970.64
#> 260 1237.104512  571130.72
#> 261 1241.344384  502823.36
#> 262 1245.663552  363433.83
#> 263 1250.177088  237005.00
#> 264 1254.943488  163738.67
#> 265 1259.774208  118738.55
#> 266 1264.692480   93834.54
#> 267 1269.610368   86969.96
#> 268 1274.428608   89435.50
#> 269 1279.245568   89842.71
#> 270 1284.078080   89889.18
#> 271 1289.020992   82225.89
#> 272 1294.008896   75893.42
#> 273 1298.990528   79939.85
#> 274 1303.907840   79126.74
#> 275 1308.854848   74153.28
#> 276 1313.883200   70790.75
#> 277 1318.965440   64867.44
#> 278 1324.078528   67890.40
#> 279 1329.110272   65083.56
#> 280 1334.155392   66187.84
#> 281 1339.165184   69974.48
#> 282 1344.014784   85475.09
#> 283 1348.760576  100035.66
#> 284 1353.494400   94283.05
#> 285 1358.352960   97629.96
#> 286 1363.126656  100834.50
#> 287 1367.891712  106282.52
#> 288 1372.606912  104146.95
#> 289 1377.472768   93014.53
#> 290 1382.423168   84035.85
#> 291 1387.471424   76629.39
#> 292 1392.506304   80515.08
#> 293 1397.386368   86482.19
#> 294 1402.188544  105214.35
#> 295 1406.864768  119586.23
#> 296 1411.561728  143193.68
#> 297 1416.113088  177644.35
#> 298 1420.623424  209716.79
#> 299 1425.043712  249562.83
#> 300 1429.462912  248299.63
#> 301 1433.947072  181043.10
#> 302 1438.631424  128547.81
#> 303 1443.586688   96348.11
#> 304 1448.628928   76418.32
#> 305 1453.737280   69520.40
#> 306 1458.826880   68627.47
#> 307 1463.864064   65342.22
#> 308 1468.958464   61326.62
#> 309 1474.166528   56428.00
#> 310 1479.382336   58858.38
#> 311 1484.397696   74146.82
#> 312 1489.139264  101209.29
#> 313 1493.846464   98255.50
#> 314 1498.592192   97132.63
#> 315 1503.226304  131607.96
#> 316 1507.656960  220932.45
#> 317 1511.979200  288177.21
#> 318 1516.312128  293626.26
#> 319 1520.762624  229780.57
#> 320 1525.368192  169128.93
#> 321 1530.174208  127555.45
#> 322 1534.997568  117470.15
#> 323 1539.705344  124688.76
#> 324 1544.320384  127689.88
#> 325 1548.939328  118750.18
#> 326 1553.641408  106842.19
#> 327 1558.411200   99105.42
#> 328 1563.211520   92877.74
#> 329 1568.045824   90089.79
#> 330 1572.960896   81551.09
#> 331 1577.970304   73784.28
#> 332 1583.039872   69485.72
#> 333 1588.125056   73041.21
#> 334 1593.152000   72746.95
#> 335 1598.226624   66765.63
#> 336 1603.331392   68972.69
#> 337 1608.307392   77605.89
#> 338 1613.239424   79105.86
#> 339 1618.179840   77178.08
#> 340 1623.217856   71822.65
#> 341 1628.399552   63187.94
#> 342 1633.660800   56019.68
#> 343 1639.017600   51608.17
#> 344 1644.399296   49953.99
#> 345 1649.681344   51560.96
#> 346 1654.868928   55412.32
#> 347 1660.028864   57153.98
#> 348 1665.218816   54987.40
#> 349 1670.488640   52958.18
#> 350 1675.802816   52925.93
#> 351 1680.859392   70226.17
#> 352 1685.790976  112569.51
#> 353 1690.332288  133683.21
#> 354 1694.946176  112808.77
#> 355 1699.745024   93473.55
#> 356 1704.781440   79132.17
#> 357 1709.966528   68089.05
#> 358 1715.188352   59105.44
#> 359 1720.502976   53649.98
#> 360 1725.868608   51968.74
#> 361 1731.182272   50193.16
#> 362 1736.428352   55790.22
#> 363 1741.454528   69985.95
#> 364 1746.213952   87195.50
#> 365 1750.948928   94648.71
#> 366 1755.732608   88965.60
#> 367 1760.735872   74481.14
#> 368 1765.846016   67360.13
#> 369 1770.933568   68480.67
#> 370 1775.933440   72880.58
#> 371 1780.902080   69665.74
#> 372 1786.037696   60773.73
#> 373 1791.257856   57196.44
#> 374 1796.504448   56241.68
#> 375 1801.729600   55297.23
#> 376 1806.997952   53105.75
#> 377 1812.321664   50604.46
#> 378 1817.679872   50666.89
#> 379 1823.005824   49574.78
#> 380 1828.365888   47099.64
#> 381 1833.787648   44543.61
#> 382 1839.229760   45154.15
#> 383 1844.647680   43889.90
#> 384 1850.023680   47066.58
#> 385 1855.318080   51694.69
#> 386 1860.535296   52791.32
#> 387 1865.753152   53021.01
#> 388 1871.091392   50829.12
#> 389 1876.459456   47097.64
#> 390 1881.819328   46529.53
#> 391 1887.179328   48361.19
#> 392 1892.503744   52287.59
#> 393 1897.709248   54767.72
#> 394 1902.676736   74971.42
#> 395 1907.315264  120708.31
#> 396 1911.743872  171743.76
#> 397 1916.130304  188038.44
#> 398 1920.580032  158103.37
#> 399 1925.176512  120590.17
#> 400 1930.067904   95429.23
#> 401 1935.084800   79992.08
#> 402 1940.096768   77472.86
#> 403 1944.897344   97938.24
#> 404 1949.482880  131764.31
#> 405 1953.959936  171140.19
#> 406 1958.399744  174330.48
#> 407 1962.865728  148755.59
#> 408 1967.504576  120372.43
#> 409 1972.365184   96508.89
#> 410 1977.502656   75660.97
#> 411 1982.746432   63695.65
#> 412 1988.050304   56092.49
#> 413 1993.442368   50545.28
#> 414 1998.944448   47205.73
#> 415 2004.452544   44860.66
#> 416 2009.992448   44287.09
#> 417 2015.582592   42160.12
#> 418 2021.176384   40319.53
#> 419 2026.766528   39802.82
#> 420 2032.364416   41553.96
#> 421 2037.889984   41984.79
#> 422 2043.394432   42674.63
#> 423 2048.780160   42280.73
#> 424 2054.156160   43683.66
#> 425 2059.525952   45884.53
#> 426 2064.901888   46255.69
#> 427 2070.250048   48196.24
#> 428 2075.499840   52824.40
#> 429 2080.667520   55701.07
#> 430 2085.843648   56953.56
#> 431 2091.065728   58653.93
#> 432 2096.295680   57909.33
#> 433 2101.563520   53836.83
#> 434 2106.913856   53803.05
#> 435 2112.303552   53061.54
#> 436 2117.655488   51119.48
#> 437 2123.001664   49673.19
#> 438 2128.401536   50866.91
#> 439 2133.789696   49543.61
#> 440 2139.199744   46327.85
#> 441 2144.597568   50083.54
#> 442 2149.977344   48566.59
#> 443 2155.339904   49612.89
#> 444 2160.673472   51334.26
#> 445 2165.963328   52631.63
#> 446 2171.185024   58052.35
#> 447 2176.326912   57872.89
#> 448 2181.537152   59269.31
#> 449 2186.721088   59294.38
#> 450 2191.921152   59824.56
#> 451 2197.168832   56013.63
#> 452 2202.448960   55429.93
#> 453 2207.704896   54326.18
#> 454 2213.013184   53780.77
#> 455 2218.282688   53959.11
#> 456 2223.607104   53338.91
#> 457 2228.956736   51465.30
#> 458 2234.353088   48113.32
#> 459 2239.734976   48998.44
#> 460 2245.078784   51544.33
#> 461 2250.402880   51855.24
#> 462 2255.678720   53517.06
#> 463 2260.910528   56701.53
#> 464 2266.052608   61549.08
#> 465 2271.104576   64751.36
#> 466 2276.212288   60576.05
#> 467 2281.462464   57586.75
#> 468 2286.736640   53824.03
#> 469 2292.052544   53633.42
#> 470 2297.382464   53827.93
#> 471 2302.672384   51146.82
#> 472 2308.042112   50045.59
#> 473 2313.460160   48223.01
#> 474 2318.864576   48136.28
#> 475 2324.234240   48382.84
#> 476 2329.614208   49025.15
#> 477 2335.008192   47647.04
#> 478 2340.380096   48420.28
#> 479 2345.713920   49109.25
#> 480 2351.123904   49113.98
#> 481 2356.486208   48629.46
#> 482 2361.839936   50192.56
#> 483 2367.173888   50917.32
#> 484 2372.549632   50875.61
#> 485 2377.865984   50236.55
#> 486 2383.201984   51605.79
#> 487 2388.545600   49659.57
#> 488 2393.873280   50375.32
#> 489 2399.221760   48904.91
#> 490 2404.609792   49640.92
#> 491 2409.951872   50876.49
#> 492 2415.249536   51973.13
#> 493 2420.561536   52518.55
#> 494 2425.838400   52670.33
#> 495 2431.067456   54755.41
#> 496 2436.263424   57481.45
#> 497 2441.431552   59280.31
#> 498 2446.569152   61695.29
#> 499 2451.675648   63589.96
#> 500 2456.763200   65969.52
#> 501 2461.866752   64545.25
#> 502 2466.994240   66188.26
#> 503 2472.055744   65987.34
#> 504 2477.135040   64439.42
#> 505 2482.234304   64558.88
#> 506 2487.284160   68127.06
#> 507 2492.304640   74201.22
#> 508 2497.270400   72559.68
#> 509 2502.285568   70325.57
#> 510 2507.288960   75465.27
#> 511 2512.226752   76877.52
#> 512 2517.211968   78495.28
#> 513 2522.157568   76341.78
#> 514 2527.193024   71782.06
#> 515 2532.219520   69855.14
#> 516 2537.258816   70942.81
#> 517 2542.292224   72988.02
#> 518 2547.242112   78483.63
#> 519 2552.132544   84096.23
#> 520 2556.910912   97727.27
#> 521 2561.682304   88240.65
#> 522 2566.507712   85610.78
#> 523 2571.380992   87045.17
#> 524 2576.256896   88827.44
#> 525 2581.099968   89448.52
#> 526 2585.945920   87493.24
#> 527 2590.793792   93197.93
#> 528 2595.486208  116162.53
#> 529 2600.101504  111747.28
#> 530 2604.850560  102755.14
#> 531 2609.658880  104397.55
#> 532 2614.436992  105277.42
#> 533 2619.196032  102747.93
#> 534 2623.947712  102507.18
#> 535 2628.721728  101306.60
#> 536 2633.503360   97588.29
#> 537 2638.261568   96801.49
#> 538 2643.053696   91817.66
#> 539 2647.770368   95570.80
#> 540 2652.392768  121175.50
#> 541 2656.907456  130090.68
#> 542 2661.485952  122954.47
#> 543 2666.186432  109676.13
#> 544 2670.950080  105662.69
#> 545 2675.669824  104045.63
#> 546 2680.356224  113485.65
#> 547 2684.980608  117417.01
#> 548 2689.625152  116220.74
#> 549 2694.281152  114280.44
#> 550 2698.993408  109010.27
#>