Chromatographic data backends

ChromBackend is a virtual class that defines what different backends need to provide to be used by the Chromatograms package and classes.

The backend should provide access to the chromatographic data which mainly consists of (paired) intensity and retention time values. Additional chromatographic metadata such as MS level and precursor and product m/z should also be provided.

The core chromatogram variables are:

chromIndex: an integer with the index of the chromatogram in the original source file (e.g. mzML file). This is a required variable for the ChromBackendMzR backend but might not be needed or defined for other backends.

collisionEnergy: for SRM data, numeric with the collision energy of the precursor.

dataOrigin: optional character with the origin of a chromatogram.

dataStorage: character defining where the data is (currently) stored.

intensity: NumericList with the intensity values of each chromatogram.

msLevel: integer defining the MS level of the data.

mz: optional numeric with the (target) m/z value for the chromatographic data.

mzMin: optional numeric with the minimal m/z value in case the data was extracted from a Spectra object.

mzMax: optional numeric with the maximal m/z value.

precursorMz: for SRM data, numeric with the target m/z of the precursor (parent).

precursorMzMin: for SRM data, optional numeric with the lower m/z of the precursor's isolation window.

precursorMzMax: for SRM data, optional numeric with the upper m/z of the precursor's isolation window.

productMz for SRM data, numeric with the target m/z of the product ion.

productMzMin: for SRM data, optional numeric with the lower m/z of the product's isolation window.

productMzMax: for SRM data, optional numeric with the upper m/z of the product's isolation window.

rtime: NumericList with the retention times of each chromatogram.

# S4 method for ChromBackend
backendInitialize(object, ...)

# S4 method for list
backendMerge(object, ...)

# S4 method for ChromBackend
backendMerge(object, ...)

# S4 method for ChromBackend
chromData(object,
  columns = chromVariables(object))

# S4 method for ChromBackend
chromData(object) <- value

# S4 method for ChromBackend
chromNames(object)

# S4 method for ChromBackend
chromNames(object) <- value

# S4 method for ChromBackend
chromIndex(object,
  columns = chromVariables(object))

# S4 method for ChromBackend
chromVariables(object)

# S4 method for ChromBackend
collisionEnergy(object)

# S4 method for ChromBackend
collisionEnergy(object) <- value

# S4 method for ChromBackend
dataOrigin(object)

# S4 method for ChromBackend
dataOrigin(object) <- value

# S4 method for ChromBackend
dataStorage(object)

# S4 method for ChromBackend
dataStorage(object) <- value

# S4 method for ChromBackend
filterDataOrigin(object, dataOrigin, ...)

# S4 method for ChromBackend
filterDataStorage(object, dataStorage, ...)

# S4 method for ChromBackend
filterMsLevel(object, msLevel)

# S4 method for ChromBackend
filterMz(object, mz, msLevel, ...)

# S4 method for ChromBackend
filterPrecursorMz(object, mz, ...)

# S4 method for ChromBackend
filterProductMz(object, mz, ...)

# S4 method for ChromBackend
intensity(object)

# S4 method for ChromBackend
intensity(object) <- value

# S4 method for ChromBackend
isEmpty(x)

# S4 method for ChromBackend
isReadOnly(object)

# S4 method for ChromBackend
length(x)

# S4 method for ChromBackend
msLevel(object)

# S4 method for ChromBackend
msLevel(object) <- value

# S4 method for ChromBackend
mz(object)

# S4 method for ChromBackend
mz(object) <- value

# S4 method for ChromBackend
mzMax(object)

# S4 method for ChromBackend
mzMax(object) <- value

# S4 method for ChromBackend
mzMin(object)

# S4 method for ChromBackend
mzMin(object) <- value

# S4 method for ChromBackend
as.list(x, ...)

# S4 method for ChromBackend
lengths(x)

# S4 method for ChromBackend
precursorMz(object)

# S4 method for ChromBackend
precursorMz(object) <- value

# S4 method for ChromBackend
precursorMzMin(object)

# S4 method for ChromBackend
precursorMzMin(object) <- value

# S4 method for ChromBackend
precursorMzMax(object)

# S4 method for ChromBackend
precursorMzMax(object) <- value

# S4 method for ChromBackend
productMz(object)

# S4 method for ChromBackend
productMz(object) <- value

# S4 method for ChromBackend
productMzMin(object)

# S4 method for ChromBackend
productMzMin(object) <- value

# S4 method for ChromBackend
productMzMax(object)

# S4 method for ChromBackend
productMzMax(object) <- value

# S4 method for ChromBackend
rtime(object)

# S4 method for ChromBackend
rtime(object) <- value

# S4 method for ChromBackend
selectChromVariables(object,
  chromVariables = chromVariables(object))

# S4 method for ChromBackend,ANY
split(x, f, drop = FALSE, ...)

# S4 method for ChromBackend
[(x, i, j, ..., drop = FALSE)

# S4 method for ChromBackend
$(x, name)

# S4 method for ChromBackend
$(x, name) <- value

ChromBackendDataFrame()

ChromBackendMzR()

Arguments

object
Object extending ChromBackend.

...
Additional arguments.

columns
For chromData accessor: optional character with column names (chromatogram variables) that should be included in the returned DataFrame. By default, all columns are returned.

value
replacement value for <- methods. See individual method description or expected data type.

dataOrigin
For filterDataOrigin: character to define which chromatograms to keep.

dataStorage
For filterDataStorage: character to define which chromatograms to keep.

msLevel
integer defining the MS level of the chromatograms to which the function should be applied. For filterMsLevel: the MS level to which object should be subsetted.

mz
For filterMz, filterPrecursorMz and filterProductMz: numeric(2) defining the lower and upper m/z of the range to subset object. Chromatograms with their min m/z (mzMin, precursorMzMin and productMz, respectively) or max m/z within mz are retained (filtering thus returns chromatograms overlapping the specified m/z range).

x
Object extending ChromBackend.

chromVariables
For selectChromVariables: character with the names of the chromatogram variables to which the backend should be subsetted.

f
factor defining the grouping to split x. See split().

drop
For [: not considered.

i
For [: integer, logical or character to subset the object.

j
For [: ignored.

name
For $ and $<-: the name of the chromatogram variable to return or set.

Backend functions

New backend classes must extend the base ChromBackend class and have to implement the following methods:

[: subset the backend. Only subsetting by element (row/i) is allowed

$, $<-: access or set/add a single chromatogram variable (column) in the backend.

as.list: returns a list of matrices with the data pairs from each chromatogram. The length of the list is equal to the number of chromatograms in object. Each element is a two-column numeric matrix with the retention time (first column) and intensity values (second column) of one chromatogram. For an empty chromatogram a matrix with 0 rows and two columns (named "rtime" and "intensity") has to be returned.

backendInitialize: initialises the backend. This method is supposed to be called right after creating an instance of the backend class and should prepare the backend. This method has to ensure to set the spectra variable dataStorage correctly.

backendMerge: merges (combines) ChromBackend objects into a single instance. All objects to be merged have to be of the same type.

chromData, chromData<-: gets or sets general chromatogram metadata (annotation, also called header). chromData returns a DataFrame, chromData<- expects a DataFrame with the same number of rows as there are chromatograms in object.

chromIndex: returns an integer vector with the index of the chromatograms in the original source file.

chromNames, chromNames<-: gets or sets the names for the chromatograms.

chromVariables: returns a character vector with the available chromatogram variables available in object.

collisionEnergy, collisionEnergy<-: gets or sets the collision energy for the precursor (for SRM data). collisionEnergy returns a numeric of length equal to the number of chromatograms in object.

dataOrigin: gets a character of length equal to the number of chromatograms in object with the data origin of each. This could e.g. be the mzML file from which the data was read.

dataStorage: gets a character of length equal to the number of chromatograms in object with the data storage of each. Note that a dataStorage of NA_character_ is not supported.

filterDataOrigin: filters the object retaining chromatograms matching any of the provided dataOrigin. Parameter dataOrigin has to be of type character and needs to match exactly the data origin value of the chromatograms to subset. filterDataOrigin should return the data ordered by the provided dataOrigin parameter, i.e. if dataOrigin = c("2", "1") was provided, the chromatograms in the resulting object should be ordered accordingly (first chromatogram from data origin "2" and then from "1").

filterDataStorage: filters the object retaining chromatograms matching any of the provided dataStorage. Parameter dataStorage has to be of type character and needs to match exactly the data storage value of the chromatograms to subset. filterDataStorage should return the data ordered by the provided dataStorage parameter, i.e. if dataStorage = c("2", "1") was provided, the chromatograms in the resulting object should be ordered accordingly (first chromatogram from data storage "2" and then from "1").

filterMsLevel: retains chromatograms of MS level msLevel.

filterMz: retains chromatograms with mzMin or mzMax within the provided m/z range (i.e. mzMax(object) >= mz[1] & mzMin(object) <= mz[2] ).

filterPrecursorMz: retains chromatograms with a their precursor m/z window overlapping the provided m/z range mz (i.e. precursorMzMax(object) >= mz[1] & precursorMzMin(object) <= mz[2]).

filterProductMz: retains chromatograms with a their product m/z window overlapping the provided m/z range mz (i.e. productMzMax(object) >= mz[1] & productMzMax(object) <= mz[2]).

intensity: gets the intensity values from the chromatograms. Returns a NumericList() of numeric vectors (intensity values for each chromatogram). The length of the list is equal to the number of chromatograms in object.

intensity<-: replaces the intensity values. value has to be a list (or NumericList()) of length equal to the number of chromatograms and the number of values within each list element identical to the number of data pairs in each chromatogram. Note that just writeable backends support this method.

isEmpty: returns a logical of length equal to the number of chromatograms with TRUE for chromatograms without any data pairs.

isReadOnly: returns a logical(1) whether the backend is read only or does allow also to write/update data.

length: returns the number of chromatograms in the object.

lengths: returns the number of data pairs (rtime-intensity values) per chromatogram.

msLevel: gets the chromatogram's MS level. Returns an integer vector (of length equal to the number of chromatograms) with the MS level for each chromatogram (or NA_integer_ if not available).

mz,mz<-: gets or sets the m/z value of the chromatograms. mz returns a numeric of length equal to the number of chromatograms in object, mz<- expects a numeric of length length(object).

mzMax,mzMax<-: gets or sets the upper m/z of the mass-to-charge range from which the chromatogram contains signal (e.g. if the chromatogram was extracted from MS data in spectra format and a m/z range was provided). mzMax returns a numeric of length equal to the number of chromatograms in object, mzMax expects a numeric of length equal to the number of chromatograms in object.

mzMin,mzMin<-: gets or sets the lower m/z of the mass-to-charge range from which the chromatogram contains signal (e.g. if the chromatogram was extracted from MS data in spectra format and a m/z range was provided). mzMin returns a numeric of length equal to the number of chromatograms in object, mzMin expects a numeric of length equal to the number of chromatograms in object.

precursorMz,precursorMz<-: gets or sets the (target) m/z of the precursor (for SRM data). precursorMz returns a numeric of length equal to the number of chromatograms in object. precursorMz<- expects a numeric of length equal to the number of chromatograms.

productMz,productMz<-: gets or sets the (target) m/z of the product (for SRM data). productMz returns a numeric of length equal to the number of chromatograms in object. productMz<- expects a numeric of length equal to the number of chromatograms.

precursorMzMin,precursorMzMax,productMzMin, productMzMax: get the lower and upper margin for the precursor or product isolation windows. These functions might return the value of productMz if the respective minimal or maximal m/z values are not defined in object.

rtime: gets the retention times from the chromatograms. returns a NumericList() of numeric vectors (retention times for each chromatogram). The length of the returned list is equal to the number of chromatograms in object.

rtime<-: replaces the retention times. value has to be a list (or NumericList()) of length equal to the number of chromatograms and the number of values within each list element identical to the number of data pairs in each chromatogram. Note that just writeable backends support this method.

selectChromVariables: reduce object retaining only specified chromatogram variables.

split: split the backend into a list of backends (depending on parameter f). The default method for ChromBackend uses split.default(), thus backends extending ChromBackend don't necessarily need to implement this method.

ChromBackendDataFrame, in-memory chromatographic data backend

The ChromBackendDataFrame objects keep all chromatographic data in memory. To reduce memory requirement, all chromatogram variables with a single value (e.g. if all are from MS level 1) are internally represented as an Rle() object that are converted into the original class (e.g. integer) when the column is accessed.

New objects can be created with the ChromBackendDataFrame() function. The backend can be subsequently initialized with the backendInitialize method, taking a DataFrame with the chromatographic data as parameter. Suggested columns of this DataFrame are:

"msLevel": integer with MS levels of the spectra.

"mz": numeric with (target) m/z of the spectra.

"mzMin","mzMax": numeric with the lower and upper m/z per chromatogram.

"precursorMz","precursorMzMin","precursorMzMax": m/z value (target, lower and upper limit) of the precursor ion (for SRM data).

"productMz","productMzMin","productMzMax": m/z value (target, lower and upper limit) of the product ion (for SRM data).

"dataOrigin": character defining the data origin.

"dataStorage": character indicating grouping of spectra in different e.g. input files. Note that missing values are not supported.

"precursorMz": numeric with the m/z value of the precursor.

"rtime": NumericList() of numeric vectors representing the retention times for the data values in a chromatogram.

"intensity": NumericList() of numeric vectors representing the intensity values for each chromatogram.

Additional columns are allowed too.

ChromBackendMzR, on-disc chromatographic data backend

The ChromBackendMzR keeps only a limited amount of data in memory, while the chromatographic data (retention time and intensity values) are fetched from the raw files on-demand. This backend uses the mzR package for data import and retrieval and hence requires that package to be installed. Also, it can only be used to import and represent data stored in mzML files.

The ChromBackendMzR backend extends the ChromBackendDataFrame backend using its DataFrame to keep chromatogram variables (except retention time and intensity) in memory.

New objects can be created with the ChromBackendMzR() function which can be subsequently filled with data by calling backendInitialize passing the file names of the input data files with argument files.

Implementation notes

Backends extending ChromBackend must implement all of its methods (listed above). Developers of new ChromBackends should follow the ChromBackendDataFrame implementation.

The ChromBackend defines the following slots:

@readonly: logical(1) whether the backend supports writing/replacing of m/z or intensity values.

Examples

## In-memory backend

## Create a simple data set with 3 chromatograms
df <- data.frame(msLevel = c(1L, 2L, 1L),
    mz = c(34.5, 454.2, 123.4))
df$rtime <- list(1:4, 2:3, 1:5)
df$intensity <- list(c(34, 25, 5, 1), c(5, 2), c(1, 5, 7, 4, 2))

be <- backendInitialize(ChromBackendDataFrame(), df)
be

#> ChromBackendDataFrame with 3 chromatograms
#>     msLevel        mz chromIndex
#>   <integer> <numeric>  <integer>
#> 1         1      34.5         NA
#> 2         2     454.2         NA
#> 3         1     123.4         NA
#>  ... 13 more variables/columns.

## Access variables
msLevel(be)

#> [1] 1 2 1

be$msLevel

#> [1] 1 2 1

rtime(be)

#> NumericList of length 3
#> [[1]] 1 2 3 4
#> [[2]] 2 3
#> [[3]] 1 2 3 4 5

be$rtime

#> NumericList of length 3
#> [[1]] 1 2 3 4
#> [[2]] 2 3
#> [[3]] 1 2 3 4 5

## Available variables
chromVariables(be)

#>  [1] "chromIndex"      "collisionEnergy" "dataOrigin"      "dataStorage"    
#>  [5] "intensity"       "msLevel"         "mz"              "mzMin"          
#>  [9] "mzMax"           "precursorMz"     "precursorMzMin"  "precursorMzMax" 
#> [13] "productMz"       "productMzMin"    "productMzMax"    "rtime"          

## Add an additional variable
be$new_var <- c("a", "b", "c")

chromVariables(be)

#>  [1] "chromIndex"      "collisionEnergy" "dataOrigin"      "dataStorage"    
#>  [5] "intensity"       "msLevel"         "mz"              "mzMin"          
#>  [9] "mzMax"           "precursorMz"     "precursorMzMin"  "precursorMzMax" 
#> [13] "productMz"       "productMzMin"    "productMzMax"    "rtime"          
#> [17] "new_var"        

be$new_var

#> [1] "a" "b" "c"

## Extract the full data
chromData(be)

#> DataFrame with 3 rows and 17 columns
#>   chromIndex collisionEnergy  dataOrigin dataStorage     intensity   msLevel
#>    <integer>       <numeric> <character> <character> <NumericList> <integer>
#> 1         NA              NA          NA    <memory>   34,25,5,...         1
#> 2         NA              NA          NA    <memory>           5,2         2
#> 3         NA              NA          NA    <memory>     1,5,7,...         1
#>          mz     mzMin     mzMax precursorMz precursorMzMin precursorMzMax
#>   <numeric> <numeric> <numeric>   <numeric>      <numeric>      <numeric>
#> 1      34.5        NA        NA          NA             NA             NA
#> 2     454.2        NA        NA          NA             NA             NA
#> 3     123.4        NA        NA          NA             NA             NA
#>   productMz productMzMin productMzMax         rtime     new_var
#>   <numeric>    <numeric>    <numeric> <NumericList> <character>
#> 1        NA           NA           NA     1,2,3,...           a
#> 2        NA           NA           NA           2,3           b
#> 3        NA           NA           NA     1,2,3,...           c