The Chromatograms class to manage and access chromatographic data

The Chromatograms class encapsules chromatographic data and related metadata. The chromatographic data is represented by a backend extending the virtual ChromBackend class which provides the raw data to the Chromatograms object. Different backends and their properties are explained in the ChromBackend class documentation.

Chromatograms(backend = ChromBackendMemory(), processingQueue = list(), ...)

Arguments

backend

ChromBackend object providing the raw data for the Chromatograms object.

processingQueue

list a list of processing steps to be applied to the chromatograms data. Each element in the list is a function that processes the chromatograms data. The processing steps are applied in the order they are listed in the processingQueue.

...

Additional arguments.

Data stored in a Chromatograms object

The Chromatograms object is a container for chromatographic data that includes peaks data (retention time and related intensity values, also referred to as peaks data variables in the context of Chromatograms) and metadata of individual chromatogram (so called chromatograms variables). While a core set of chromatograms variables (the coreChromatogramsVariables()) and peaks data variables (the corePeaksVariables()) are guaranteed to be provided by a Chromatograms, it is possible to add arbitrary additional chromatograms variables to a Chromatograms object.

The chromatograms variables information can be accessed using the chromData() function. it is also possible to access specific chromatograms variables using $. chromData can be accessed, replaced but also filtered/subsetted. Refer t the chromData documentation for more details.

The Chromatograms object is designed to contain chromatographic data of a (large) set of chromatograms. The data is organized linearly and can be thought of a list of chromatograms, i.e. each element in the Chromatograms is one chromatogram.

Creation of objects

Chromatograms objects can be created using the Chromatograms() construction function.

See also

Examples

## create a Chromatograms object
chroms <- Chromatograms(backend = ChromBackendMemory())